Add ability to parse ref/alt alleles from HGVS nomenclature
Sometimes the ref/alt allele fields are not available in the description
column. This is either because they are not offered or because they
simply have a value of 'N/A'. This commit adds the ability to recover
the ref/alt allele(s) from the HGVS nomenclature. In fact, it will check
for the HGVS field first. If that's missing, only then will it check for
the ref and alt fields.