Refactor and improve variant query
Much of the code for querying by variant was refactored in order to use
one class for all sources rather than multiple classes for multiple
sources. The process was also improved by reducing the complexity and
making it easier for the developer to add another source to query by
variant.